Make tile source always the same
This will make it easier to swap tile sets out. Instead of needing to
edit this file, we can just upload new tiles to the tiles/ directory.
Instead of using the file to preserve metadata about which set of tiles
it is, we can upload a file to that directory with that information.